// types.ts
export interface DbTrainer {
  id: number;
  name: string;
  role: string;
  description: string;
  image_url: string;
  created_at?: string;
}

export interface DbOnlineCourse {
  id: number;
  title: string;
  price: string;
  description: string;
  video_url?: string;
  image_url?: string;
  // Joined Data
  online_course_trainers?: {
    trainers: DbTrainer;
  }[];
}

export interface DbFreeCourse {
  id: number;
  title: string;
  description: string;
  trainer_name: string;
  video_url: string;
  image_url?: string;
  created_at?: string;
}

export interface DbOtherProgram {
  id: number;
  title: string;
  price: string;
  description: string;
  video_url?: string;
  image_url?: string;
  trainers_count?: string;
  available_dates?: string[];
  // Joined Data
  other_program_trainers?: {
    trainers: DbTrainer;
  }[];
}

export interface DbMiniProgram {
  id: number;
  title: string;
  subtitle?: string;
  price: string;
  description?: string;
  video_url?: string;
  image_url?: string;
  learn_list?: string[];
  receive_list?: string[];
  available_dates?: string[];
  // Joined Data
  mini_program_trainers?: {
    trainers: DbTrainer;
  }[];
}

export interface DbZellPlusContent {
  id: number;
  title: string;
  description?: string;
  image_url?: string;
  video_url?: string;
  created_at?: string;
}

export interface DbPodcast {
  id: number;
  title: string;
  description?: string;
  image_url?: string;
  video_url?: string;
  created_at?: string;
}